Fix build errors on a real Qt 6 host
Neither file compiled against Qt 6.9.3. Both sit in code paths the
sandbox could never build, having no Qt installed.

runtime_agent.cpp: bytesRead and bytesWritten passed the raw transferred
count into a QJsonObject initializer. QJsonValue has no unambiguous
conversion from ssize_t, so both sites failed to compile. Cast to qint64.

render_probe.cpp: the guard applied ! to
functions->initializeOpenGLFunctions(). QOpenGLContext::functions()
returns an unversioned QOpenGLFunctions, whose initializeOpenGLFunctions()
returns void, so negating it is ill-formed. It never returned false; it
returns nothing. The bool-returning method belongs to
QAbstractOpenGLFunctions, base of the versioned QOpenGLFunctions_X_Y
classes, which is the API the original guard was written against. Resolve
QOpenGLFunctions_4_5_Core through QOpenGLVersionFunctionsFactory::get and
keep the status check, so a failed resolution is actually reported.

That pins the snippet to a 4.5-core-or-later context. main.cpp requests
3.3 core; Mesa 25.2.8 grants 4.6 core for that request, verified with an
offscreen context using main.cpp's exact QSurfaceFormat. It would fail on
a driver that grants exactly the requested version, and on macOS, where
core profiles stop at 4.1.

Verified by compiling snippets/render_probe.cpp with the project's own
flags from compile_commands.json (-O3 -Wall -Wextra -Wpedantic, clean).

The runtime_agent.cpp change was taken from the extracted archive's
working tree, not authored here.

+++ src/agent/runtime_agent.cpp
@@ -825,7 +825,7 @@ void RuntimeAgent::dispatchRequest(QLocalSocket* socket,
         data.truncate(static_cast<qsizetype>(transferred));
         sendSuccess(socket, requestId, QJsonObject{
             {QStringLiteral("address"), pointerString(reinterpret_cast<void*>(address))},
-            {QStringLiteral("bytesRead"), transferred},
+            {QStringLiteral("bytesRead"), static_cast<qint64>(transferred)},
             {QStringLiteral("base64"), QString::fromLatin1(data.toBase64())},
         });
         return;
@@ -856,7 +856,7 @@ void RuntimeAgent::dispatchRequest(QLocalSocket* socket,
         }
         sendSuccess(socket, requestId, QJsonObject{
             {QStringLiteral("address"), pointerString(reinterpret_cast<void*>(address))},
-            {QStringLiteral("bytesWritten"), transferred},
+            {QStringLiteral("bytesWritten"), static_cast<qint64>(transferred)},
         });
         return;
     }
